refactoring validation functions in the mw histmaker
