+++ scripts/histmakers/mw_with_mu_eta_pt.py
@@ -222,7 +222,7 @@ def build_graph(df, dataset):
         df = df.Filter("wrem::hasMatchDR2(goodMuons_eta0,goodMuons_phi0,GenPart_eta[postFSRmuons],GenPart_phi[postFSRmuons],0.09)")
 
     if isW or isZ:
-        df = muon_calibration.define_cvh_reco_muon_kinematics(df)
+        df = muon_validation.define_cvh_reco_muon_kinematics(df)
         #FIXME: make the smearing weights work without filtering on taus
         if args.muonScaleVariation == 'smearingWeights':
             reco_sel = "vetoMuonsPre"
@@ -248,15 +248,9 @@ def build_graph(df, dataset):
                 df = df.Define(f"{reco_sel_GF}_{var.lower()}0_gen_smeared", f"{reco_sel_GF}_genSmeared{var.capitalize()}[0]")
             df = df.Define(f"{reco_sel_GF}_covMat0", f"{reco_sel_GF}_covMat[0]")
 
-#            reco_sel = "vetoMuonsPre"
-#            df = muon_calibration.define_matched_gen_muons_kinematics(df, reco_sel)
-#            df = muon_calibration.calculate_matched_gen_muon_kinematics(df, reco_sel)
-#            df = muon_calibration.define_matched_gen_muons_covMat(df, reco_sel)
-#            df = muon_calibration.define_matched_genSmeared_muon_kinematics(df, reco_sel)
-
         if args.validationHists:
             for reco_type in ['crctd', 'cvh', 'uncrct', 'gen_smeared']:
-                df = muon_calibration.define_reco_over_gen_cols(df, reco_type)
+                df = muon_validation.define_reco_over_gen_cols(df, reco_type)
 
     df = df.Define("goodMuons_pfRelIso04_all0", "Muon_pfRelIso04_all[goodMuons][0]")
 
@@ -368,7 +362,7 @@ def build_graph(df, dataset):
         if args.muonScaleVariation == 'smearingWeights' and (isW or isZ): 
             nominal_cols_gen, nominal_cols_gen_smeared = muon_calibration.make_alt_reco_and_gen_hists(df, results, axes, cols, reco_sel_GF)
             if args.validationHists: 
-                wremnants.make_reco_over_gen_hists(df, results)
+                muon_validation.make_reco_over_gen_hists(df, results)
 
     if not dataset.is_data and not args.onlyMainHistograms:
         
@@ -573,57 +567,14 @@ def build_graph(df, dataset):
 
             if args.muonScaleVariation == 'smearingWeights':
                 if args.validationHists:
-                    df = wremnants.define_cols_for_smearing_weights(df, calibration_uncertainty_helper)
-                    wremnants.make_hists_for_smearing_weights(df, axes, cols, results)
-
-                df = df.Define("goodMuons_pt0_gen_smeared_scaleUp_mil", "goodMuons_pt0_gen_smeared * 1.001")
-                df = df.Define("goodMuons_pt0_gen_smeared_scaleDn_mil", "goodMuons_pt0_gen_smeared / 1.001")
-                df = df.Define("goodMuons_pt0_scaleUp_tenthmil", "goodMuons_pt0 * 1.0001")
-                df = df.Define("goodMuons_pt0_scaleDn_tenthmil", "goodMuons_pt0 / 1.0001")
-                df = df.Define("goodMuons_pt0_gen_smeared_scaleUp_tenthmil", "goodMuons_pt0_gen_smeared * 1.0001")
-                df = df.Define("goodMuons_pt0_gen_smeared_scaleDn_tenthmil", "goodMuons_pt0_gen_smeared / 1.0001")
-                muonScaleVariationUpMil = df.HistoBoost(
-                    "nominal_muonScaleVariationUpMil", 
-                    axes,
-                    [nominal_cols_gen_smeared[0], "goodMuons_pt0_gen_smeared_scaleUp_mil", *nominal_cols_gen_smeared[2:], "nominal_weight"], 
-                    storage=hist.storage.Double()
-                )
-                muonScaleVariationDnMil = df.HistoBoost(
-                    "nominal_muonScaleVariationDnMil", 
-                    axes,
-                    [nominal_cols_gen_smeared[0], "goodMuons_pt0_gen_smeared_scaleDn_mil", *nominal_cols_gen_smeared[2:], "nominal_weight"], 
-                    storage=hist.storage.Double()
-                )
-                muonScaleVariationUpTenthmil = df.HistoBoost(
-                    "nominal_muonScaleVariationUpTenthmil", 
-                    axes,
-                    [cols[0], "goodMuons_pt0_scaleUp_tenthmil", *cols[2:], "nominal_weight"], 
-                    storage=hist.storage.Double()
-                )
-                muonScaleVariationDnTenthmil = df.HistoBoost(
-                    "nominal_muonScaleVariationDnTenthmil", 
-                    axes,
-                    [cols[0], "goodMuons_pt0_scaleDn_tenthmil", *cols[2:], "nominal_weight"], 
-                    storage=hist.storage.Double()
-                )
-                muonScaleVariationUpTenthmil_gen_smear = df.HistoBoost(
-                    "nominal_muonScaleVariationUpTenthmil_gen_smear", 
-                    axes,
-                    [nominal_cols_gen_smeared[0], "goodMuons_pt0_gen_smeared_scaleUp_tenthmil", *nominal_cols_gen_smeared[2:], "nominal_weight"], 
-                    storage=hist.storage.Double()
-                )
-                muonScaleVariationDnTenthmil_gen_smear = df.HistoBoost(
-                    "nominal_muonScaleVariationDnTenthmil_gen_smear", 
-                    axes,
-                    [nominal_cols_gen_smeared[0], "goodMuons_pt0_gen_smeared_scaleDn_tenthmil", *nominal_cols_gen_smeared[2:], "nominal_weight"], 
-                    storage=hist.storage.Double()
-                )
-                results.append(muonScaleVariationUpMil)
-                results.append(muonScaleVariationDnMil)
-                results.append(muonScaleVariationUpTenthmil)
-                results.append(muonScaleVariationDnTenthmil)
-                results.append(muonScaleVariationUpTenthmil_gen_smear)
-                results.append(muonScaleVariationDnTenthmil_gen_smear)
+                    df = muon_validation.define_cols_for_smearing_weights_perse(
+                        df, jpsi_crctn_data_unc_helper, reco_sel_GF
+                    )
+                    df = muon_validation.define_cols_for_manual_shifts(df)
+                    muon_validation.make_hists_for_smearing_weights_perse(df, axes, cols, results)
+                    muon_validation.make_hists_for_manual_scale_shifts(
+                        df, axes, cols, nominal_cols_gen_smeared, results
+                    )
 
             df = df.Define("Muon_cvhMomCov", "wrem::splitNestedRVec(Muon_cvhMomCov_Vals, Muon_cvhMomCov_Counts)")
 
@@ -655,6 +606,7 @@ def build_graph(df, dataset):
 resultdict = narf.build_and_run(datasets, build_graph)
 if not args.onlyMainHistograms and args.muonScaleVariation == 'smearingWeights':
     muon_calibration.transport_smearing_weights_to_reco(resultdict, nonClosureScheme = args.nonClosureScheme)
-    muon_calibration.muon_scale_variation_from_manual_shift(resultdict)
+    if args.validationHists:
+        muon_validation.muon_scale_variation_from_manual_shift(resultdict)
 
 output_tools.write_analysis_output(resultdict, f"{os.path.basename(__file__).replace('py', 'hdf5')}", args, update_name=not args.forceDefaultName)
