+++ wremnants/muon_calibration.py
@@ -537,22 +537,6 @@ def define_corrected_reco_muon_kinematics(df, muons="goodMuons", kinematic_vars 
         )
     return df
 
-def define_cvh_reco_muon_kinematics(df, kinematic_vars = ["pt", "eta", "phi", "charge"]):
-    for var in kinematic_vars:
-        df = df.Define(
-            f"goodMuons_{var.lower()}0_cvh",
-            f"Muon_cvh{var.capitalize()}[goodMuons][0]"
-        )
-    return df
-
-def define_uncrct_reco_muon_kinematics(df, kinematic_vars = ["pt", "eta", "phi", "charge"]):
-    for var in kinematic_vars:
-        df = df.Define(
-            f"goodMuons_{var.lower()}0_uncrct",
-            f"Muon_{var.lower()}[goodMuons][0]"
-        )
-    return df
-
 def transport_smearing_weights_to_reco(
     resultdict,
     procs = ['WplusmunuPostVFP', 'WminusmunuPostVFP', 'ZmumuPostVFP'],
@@ -602,14 +586,6 @@ def transport_smearing_weights_to_reco(
                 logger.warning("nuisances generated by smearing weights not transported to RECO kinematics!")
                 return
 
-def muon_scale_variation_from_manual_shift(
-    resultdict, procs = ['WplusmunuPostVFP', 'WminusmunuPostVFP', 'ZmumuPostVFP'],
-):
-    for proc in procs:
-        proc_hists = resultdict[proc]['output']
-        manual_shift_hists = [proc_hists['nominal_muonScaleVariationDnTenthmil'].get(), proc_hists['nominal_muonScaleVariationUpTenthmil'].get()]
-        proc_hists['muonScaleSyst_manualShift'] = hh.combineUpDownVarHists(*manual_shift_hists)
-
 def make_alt_reco_and_gen_hists(df, results, nominal_axes, nominal_columns, matched_reco_sel = "goodMuons"):
 
     nominal_cols_gen = nominal_columns[:]
@@ -629,75 +605,6 @@ def make_alt_reco_and_gen_hists(df, results, nominal_axes, nominal_columns, matc
 
     return [nominal_cols_gen, nominal_cols_gen_smeared]
     
-####################
-## FOR VALIDATION ##
-####################
-
-def define_reco_over_gen_cols(df, reco_type, kinematic_vars = ['pt', 'eta']):
-    kinematic_vars = common.string_to_list(kinematic_vars)
-    for var in kinematic_vars:
-        reco_col = f"goodMuons_{var.lower()}0" if reco_type == 'crctd' \
-                   else f"goodMuons_{var.lower()}0_{reco_type}"
-        df = df.Define(
-            f"goodMuons_{var.lower()}0_{reco_type}_over_gen",
-            f"{reco_col}/goodMuons_{var.lower()}0_gen"
-        )
-    return df
-
-def make_reco_over_gen_hists(df, results):
-    nominal_cols_crctd_over_gen = [
-        "goodMuons_pt0_crctd_over_gen"
-    ]
-    nominal_cols_cvh_over_gen = [
-        "goodMuons_pt0_cvh_over_gen"
-    ]
-    nominal_cols_uncrct_over_gen = [
-        "goodMuons_pt0_uncrct_over_gen"
-    ]
-    nominal_cols_gen_smeared_over_gen = [
-        "goodMuons_pt0_gen_smeared_over_gen",
-    ]
-    axis_pt_reco_over_gen = hist.axis.Regular(1000, 0.9, 1.1, underflow=True, overflow=True, name = "reco_pt_over_gen")
-    axis_qop_reco_over_gen = hist.axis.Regular(1000, 0.9, 1.1, underflow=True, overflow=True, name = "reco_qop_over_gen")
-    crctd_over_gen =  df.HistoBoost("crctd_over_gen", [axis_pt_reco_over_gen], [*nominal_cols_crctd_over_gen, "nominal_weight"], storage=hist.storage.Double())
-    cvh_over_gen =  df.HistoBoost("cvh_over_gen", [axis_pt_reco_over_gen], [*nominal_cols_cvh_over_gen, "nominal_weight"], storage=hist.storage.Double())
-    uncrct_over_gen = df.HistoBoost("uncrct_over_gen", [axis_pt_reco_over_gen], [*nominal_cols_uncrct_over_gen, "nominal_weight"], storage=hist.storage.Double())
-    gen_smeared_over_gen = df.HistoBoost("gen_smeared_over_gen", [axis_pt_reco_over_gen], [*nominal_cols_gen_smeared_over_gen, "nominal_weight"], storage=hist.storage.Double())
-    
-    results.append(crctd_over_gen)
-    results.append(cvh_over_gen)
-    results.append(uncrct_over_gen)
-    results.append(gen_smeared_over_gen)
-
-def define_cols_for_smearing_weights(df, helper_func):
-    df = df.Define("unity_as_double", "1.0")
-    df = df.Define("muonScaleSyst_smearingWeightsPerSe_tensor", helper_func,
-        [
-        "goodMuons_qop0_gen_smeared",
-        "goodMuons_pt0_gen_smeared_a_la_qop",
-        "goodMuons_eta0_gen_smeared",
-        "goodMuons_phi0_gen_smeared",
-        "goodMuons_charge0_gen_smeared",
-        "covMat_goodGenMuons0",
-        "goodMuons_qop0_gen",
-        "goodMuons_pt0_gen",
-        "goodMuons_eta0_gen",
-        "goodMuons_phi0_gen",
-        "goodMuons_charge0_gen",
-        "unity_as_double"
-        ]
-    )
-    df = df.Define("smearing_weights_down", "muonScaleSyst_smearingWeightsPerSe_tensor(0,0)")
-    df = df.Define("smearing_weights_up", "muonScaleSyst_smearingWeightsPerSe_tensor(0,1)")
-    return df
-
-def make_hists_for_smearing_weights(df, nominal_axes, nominal_cols, results):
-    axis_smearing_weight = hist.axis.Regular(1000, 0.99, 1.01, underflow=True, overflow=True, name = "smearing_weight")
-    smearing_weights_down = df.HistoBoost("smearing_weights_down", [*nominal_axes, axis_smearing_weight], [*nominal_cols, "smearing_weights_down"], storage=hist.storage.Double())
-    smearing_weights_up = df.HistoBoost("smearing_weights_up", [*nominal_axes, axis_smearing_weight], [*nominal_cols, "smearing_weights_up"], storage=hist.storage.Double())
-    results.append(smearing_weights_down)
-    results.append(smearing_weights_up)
-
 def define_lbl_corrections_jpsi_calibration_ntuples(df, helper):
     df = df.DefinePerSample("Muplus_charge", "1")
     df = df.DefinePerSample("Muminus_charge", "-1")
