+++ wremnants/muon_validation.py
@@ -5,6 +5,7 @@ import numpy as np
 import uproot
 from functools import reduce
 from utilities import common, logging
+from utilities import boostHistHelpers as hh
 from wremnants.muon_calibration import get_jpsi_scale_param_cov_mat
 
 narf.clingutils.Declare('#include "muon_validation.h"')
@@ -145,3 +146,156 @@ def define_jpsi_crctd_unc_z_mass(df):
         )
     )
     return df
+
+def define_cvh_reco_muon_kinematics(df, kinematic_vars = ["pt", "eta", "phi", "charge"]):
+    for var in kinematic_vars:
+        col_name = f"goodMuons_{var.lower()}0_cvh"
+        if not (col_name in df.GetColumnNames()):
+            df = df.Define(col_name, f"Muon_cvh{var.capitalize()}[goodMuons][0]")
+    return df
+
+def define_uncrct_reco_muon_kinematics(df, kinematic_vars = ["pt", "eta", "phi", "charge"]):
+    for var in kinematic_vars:
+        col_name = f"goodMuons_{var.lower()}0_uncrct"
+        if not (col_name in df.GetColumnNames()):
+            df = df.Define(col_name, f"Muon_{var.lower()}[goodMuons][0]")
+    return df
+
+def define_reco_over_gen_cols(df, reco_type, kinematic_vars = ['pt', 'eta']):
+    kinematic_vars = common.string_to_list(kinematic_vars)
+    df = define_cvh_reco_muon_kinematics(df, kinematic_vars)
+    df = define_uncrct_reco_muon_kinematics(df, kinematic_vars)
+    for var in kinematic_vars:
+        reco_col = f"goodMuons_{var.lower()}0" if reco_type == 'crctd' \
+                   else f"goodMuons_{var.lower()}0_{reco_type}"
+        df = df.Define(
+            f"goodMuons_{var.lower()}0_{reco_type}_over_gen",
+            f"{reco_col}/goodMuons_{var.lower()}0_gen"
+        )
+    return df
+
+def make_reco_over_gen_hists(df, results):
+    nominal_cols_crctd_over_gen = [
+        "goodMuons_pt0_crctd_over_gen"
+    ]
+    nominal_cols_cvh_over_gen = [
+        "goodMuons_pt0_cvh_over_gen"
+    ]
+    nominal_cols_uncrct_over_gen = [
+        "goodMuons_pt0_uncrct_over_gen"
+    ]
+    nominal_cols_gen_smeared_over_gen = [
+        "goodMuons_pt0_gen_smeared_over_gen",
+    ]
+    axis_pt_reco_over_gen = hist.axis.Regular(1000, 0.9, 1.1, underflow=True, overflow=True, name = "reco_pt_over_gen")
+    axis_qop_reco_over_gen = hist.axis.Regular(1000, 0.9, 1.1, underflow=True, overflow=True, name = "reco_qop_over_gen")
+    crctd_over_gen =  df.HistoBoost("crctd_over_gen", [axis_pt_reco_over_gen], [*nominal_cols_crctd_over_gen, "nominal_weight"], storage=hist.storage.Double())
+    cvh_over_gen =  df.HistoBoost("cvh_over_gen", [axis_pt_reco_over_gen], [*nominal_cols_cvh_over_gen, "nominal_weight"], storage=hist.storage.Double())
+    uncrct_over_gen = df.HistoBoost("uncrct_over_gen", [axis_pt_reco_over_gen], [*nominal_cols_uncrct_over_gen, "nominal_weight"], storage=hist.storage.Double())
+    gen_smeared_over_gen = df.HistoBoost("gen_smeared_over_gen", [axis_pt_reco_over_gen], [*nominal_cols_gen_smeared_over_gen, "nominal_weight"], storage=hist.storage.Double())
+    
+    results.append(crctd_over_gen)
+    results.append(cvh_over_gen)
+    results.append(uncrct_over_gen)
+    results.append(gen_smeared_over_gen)
+
+def define_cols_for_smearing_weights_perse(df, helper_func, reco_sel_GF):
+    if not ("unity" in df.GetColumnNames()):
+        df = df.Define("unity", "1.0")
+    if not ("bool_false" in df.GetColumnNames()):
+        df = df.DefinePerSample("bool_false", "false")
+    df = df.Define("muonScaleSyst_smearingWeightsPerSe_tensor", helper_func,
+        [
+            f"{reco_sel_GF}_genQop",
+            f"{reco_sel_GF}_genPhi",
+            f"{reco_sel_GF}_genEta",
+            f"{reco_sel_GF}_genSmearedQop",
+            f"{reco_sel_GF}_genSmearedPhi",
+            f"{reco_sel_GF}_genSmearedEta",
+            f"{reco_sel_GF}_genSmearedCharge",
+            f"{reco_sel_GF}_genSmearedPt",
+            f"{reco_sel_GF}_covMat",
+            "unity",
+            "bool_false"
+        ]
+    )
+    df = df.Define("smearing_weights_down", "muonScaleSyst_smearingWeightsPerSe_tensor(0,0)")
+    df = df.Define("smearing_weights_up", "muonScaleSyst_smearingWeightsPerSe_tensor(0,1)")
+    return df
+
+def define_cols_for_manual_shifts(df):
+    df = df.Define("goodMuons_pt0_gen_smeared_scaleUp_mil", "goodMuons_pt0_gen_smeared * 1.001")
+    df = df.Define("goodMuons_pt0_gen_smeared_scaleDn_mil", "goodMuons_pt0_gen_smeared / 1.001")
+    df = df.Define("goodMuons_pt0_scaleUp_tenthmil", "goodMuons_pt0 * 1.0001")
+    df = df.Define("goodMuons_pt0_scaleDn_tenthmil", "goodMuons_pt0 / 1.0001")
+    df = df.Define("goodMuons_pt0_gen_smeared_scaleUp_tenthmil", "goodMuons_pt0_gen_smeared * 1.0001")
+    df = df.Define("goodMuons_pt0_gen_smeared_scaleDn_tenthmil", "goodMuons_pt0_gen_smeared / 1.0001")
+    return df
+
+def make_hists_for_smearing_weights_perse(df, nominal_axes, nominal_cols, results):
+    axis_smearing_weight = hist.axis.Regular(1000, 0.99, 1.01, underflow=True, overflow=True, name = "smearing_weight")
+    smearing_weights_down = df.HistoBoost("smearing_weights_down", [*nominal_axes, axis_smearing_weight], [*nominal_cols, "smearing_weights_down"], storage=hist.storage.Double())
+    smearing_weights_up = df.HistoBoost("smearing_weights_up", [*nominal_axes, axis_smearing_weight], [*nominal_cols, "smearing_weights_up"], storage=hist.storage.Double())
+    results.append(smearing_weights_down)
+    results.append(smearing_weights_up)
+
+# make the histograms for muon momentum scale shifts by
+# manually shift the muon pt by a certain proportion
+def make_hists_for_manual_scale_shifts(df, axes, cols, cols_gen_smeared, results):
+    muonScaleVariationUpMil = df.HistoBoost(
+        "nominal_muonScaleVariationUpMil", 
+        axes,
+        [cols_gen_smeared[0], "goodMuons_pt0_gen_smeared_scaleUp_mil", *cols_gen_smeared[2:], "nominal_weight"], 
+        storage=hist.storage.Double()
+    )
+    muonScaleVariationDnMil = df.HistoBoost(
+        "nominal_muonScaleVariationDnMil", 
+        axes,
+        [cols_gen_smeared[0], "goodMuons_pt0_gen_smeared_scaleDn_mil", *cols_gen_smeared[2:], "nominal_weight"], 
+        storage=hist.storage.Double()
+    )
+    muonScaleVariationUpTenthmil = df.HistoBoost(
+        "nominal_muonScaleVariationUpTenthmil", 
+        axes,
+        [cols[0], "goodMuons_pt0_scaleUp_tenthmil", *cols[2:], "nominal_weight"], 
+        storage=hist.storage.Double()
+    )
+    muonScaleVariationDnTenthmil = df.HistoBoost(
+        "nominal_muonScaleVariationDnTenthmil", 
+        axes,
+        [cols[0], "goodMuons_pt0_scaleDn_tenthmil", *cols[2:], "nominal_weight"], 
+        storage=hist.storage.Double()
+    )
+    muonScaleVariationUpTenthmil_gen_smear = df.HistoBoost(
+        "nominal_muonScaleVariationUpTenthmil_gen_smear", 
+        axes,
+        [cols_gen_smeared[0], "goodMuons_pt0_gen_smeared_scaleUp_tenthmil", *cols_gen_smeared[2:], "nominal_weight"], 
+        storage=hist.storage.Double()
+    )
+    muonScaleVariationDnTenthmil_gen_smear = df.HistoBoost(
+        "nominal_muonScaleVariationDnTenthmil_gen_smear", 
+        axes,
+        [cols_gen_smeared[0], "goodMuons_pt0_gen_smeared_scaleDn_tenthmil", *cols_gen_smeared[2:], "nominal_weight"], 
+        storage=hist.storage.Double()
+    )
+    results.append(muonScaleVariationUpMil)
+    results.append(muonScaleVariationDnMil)
+    results.append(muonScaleVariationUpTenthmil)
+    results.append(muonScaleVariationDnTenthmil)
+    results.append(muonScaleVariationUpTenthmil_gen_smear)
+    results.append(muonScaleVariationDnTenthmil_gen_smear)
+
+def muon_scale_variation_from_manual_shift(
+    resultdict,
+    procs = ['WplusmunuPostVFP', 'WminusmunuPostVFP', 'ZmumuPostVFP'],
+):
+    for proc in procs:
+        proc_hists = resultdict[proc]['output']
+        manual_shift_hists = [
+            proc_hists['nominal_muonScaleVariationDnTenthmil'].get(),
+            proc_hists['nominal_muonScaleVariationUpTenthmil'].get()
+        ]
+        proc_hists['muonScaleSyst_manualShift'] = narf.ioutils.H5PickleProxy(
+            hh.combineUpDownVarHists(*manual_shift_hists)
+        )
+
